Move overflow check addition to helpers file.
